`Paywalls`: refactored `PurchaseHandler` extracting protocol
As suggested by @aboedo in https://github.com/RevenueCat/purchases-ios/pull/3164#discussion_r1323603537
